Build.sh: Change default build type to DebugOptimized
DebugOptimized on most platforms (macOS, Linux) takes a similar amount
of time to build as Debug, however it adds cheap-to-enable
optimizations which still allow debugging but give a significant
performance boost to the generated executables.

For example, one of the kv_engine testsuites
(ep_testsuite.value_eviction.comp_passive) takes 45s to run on an M1
Pro MacBookPro in Debug ; but that reduces to 20s under
DebugOptimized.

Given people normally run executables (at least) once after compiling,
this is generally a good trade-off. As such, change the default
CMAKE_BUILD_TYPE to DebugOptimized if not otherwise specified.

Change-Id: Iad20a3d557634bf6d463d0f19b876779c4537014
Reviewed-on: https://review.couchbase.org/c/tlm/+/194405
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>