CBD-4334: Specify CC for cgo for all linux builds
Without the change cgo will try to use gcc from path, which may not
be the same compiler used to build the rest of the source.

As part of building on the new CV images /usr/local/bin/gcc would
be 10.2.0, and if we tried to build using 13.2.0 we would get
link failures as they have a different libstdc++ library (sigar
would use the one from 13.2.0, but go would find the one from
10.2.0 which don't have the same version).

Given that we bundle the libstdc++ version in install/lib, we
would have "other" problems by having go programs being linked
with a "random" compiler from path.

Change-Id: I1ad529529b9df5d1da98a78b15edf86fc845d960
Reviewed-on: https://review.couchbase.org/c/tlm/+/197010
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>