CBD-6348: [BP] Disable '-warn_duplicate_libraries' ld option under Clang 15
AppleClang-15 enables the linker option '-warn_duplicate_libraries' by
default - "Warn if the input contains duplicate library options".
Across our tree we have various places where a given library ends up
listed twice on a linker line - for example 'libplatform' which could
be a direct and indirect dependent of a target.

Listing a library twice is benign - the linker will just include it
once in the order of the first instance, so disable this warning to
avoid the noise.

Change-Id: I94c689ea1d5d83deb9412879f2fa4dae695360ac
Reviewed-on: https://review.couchbase.org/c/tlm/+/235643
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker