MB-58833: Add gcc -Wshadow=compatible-local flag
We have recently hit a nasty issue caused by variable shadowing and the
consequent usage of the wrong variable reference (MB-58463).

The -Wshadow flag would give us a compilation warning and reduce the
chance of making that kind of error.

I start by introducing -Wshadow=compatible-local for GCC. That is the
loosest of multiple options for that flag.
That has been tested on KV. New KV warnings already fixed at
https://review.couchbase.org/c/kv_engine/+/197732.

In subsequent patches I'll make equivalent changes for Clang and MSVC.

Change-Id: I62b30da2b2311e4f0d32d6eca4c3f00a1c85bd0f
Reviewed-on: https://review.couchbase.org/c/tlm/+/197731
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>