MB-58915: Create all necessary libraries in standalone tools package
At INSTALL time, look at the installed executabes and have CMake
introspect the set of runtime library dependencies. Track those
libraries down and install them to the tools package as well.

For now this logic is embedded in GoModBuild() as the only executables
we ship are built that way. However it's mostly language-agnostic so it
could be broken out to a separate function in future.

The ALT_INSTALL_PATH argument is changed to ADD_TO_TOOLS, to consolidate
the logic in one place - now that we need to also install dependent
libraries, we need to make more assumptions about the tools package's
directory layout.

Change-Id: I01196677bb420b3006a1f3fb9f8dca4239260030
Reviewed-on: https://review.couchbase.org/c/tlm/+/201954
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>