[Re-apply] MB-60560: Upgrade rebar
rebar is deprecated and un-maintained. rebar3 is the new standard
compilation tool for erlang code and has benefits that we can take
advantage of.

Re-application of https://review.couchbase.org/c/tlm/+/202101, the
linked ns_server change hit a merge conflict. This is linked by topic to the ns_server change.

Change-Id: I915346b2326584792c6d5ed69e9d12fcbe451ebe
Reviewed-on: https://review.couchbase.org/c/tlm/+/207076
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>