CBD-5783: Ensure CXX is passed to CGO
This commit addresses an issue where CGO was defaulting to the first C++
compiler (GCC 10.2.0) found in the system's PATH when compiling C++ code. This behaviour led to an unintended linkage against `libasan.so.6` instead of the desired `libasan.so.8` (GCC 13.2.0)

By passing CXX, go references the correct GCC, and links against the
correct libraries.

Change-Id: I1b39537fd3d9d4f64e0793fddf276436283934ae
Reviewed-on: https://review.couchbase.org/c/tlm/+/204499
Tested-by: Chris Hillery <ceej@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Chris Hillery <ceej@couchbase.com>