MB-60571: Make clang-format ignore JSON files
clang-format fails on attempt to format any JSON files because is not
configured for JSON.

This patch will disable clang-format for JSON, which avoids the error.

The reason this patch disables as opposed to setting up JSON formatting
is that we have a couple of JSON configuration files in kv_engine which
are well-formatted but do not follow a specific convention.

Change-Id: Ie89e23684710ac8bbef3a318e91e0db0d041385a
Reviewed-on: https://review.couchbase.org/c/tlm/+/208581
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>