MB-60571: Make clang-format ignore JSON files
clang-format fails on attempt to format any JSON files because is not
configured for JSON.

This patch will disable clang-format for JSON, which avoids the error.

The reason this patch disables as opposed to setting up JSON formatting
is that we have a couple of JSON configuration files in kv_engine which
are well-formatted but do not follow a specific convention.

Change-Id: I4fe6326e57624f57392be1cade4c57d248b51f43
Reviewed-on: https://review.couchbase.org/c/tlm/+/208248
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>