Upgrade spdlog to 1.15.0
Upgrade spdlog to 1.15.0 as second version as the first version
was part of upgrading folly (and includes a newer fmt).

This rebuild just updates spdlog to 1.15.0 without updating
any of the dependencies. Instead of forcing use of clang
to build the TSAN version of the library we'll be using
the default compiler (gcc 13.2.0 on Linux)

Change-Id: I0590db3f45e3d9d6f0c8b42583d9dcff3de9fb9d
Reviewed-on: https://review.couchbase.org/c/tlm/+/221826
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>