CBD-6269: Introduce GoModTidySetup() macro
Previously we introduced `-tidy` targets as a side-effect of
`GoModBuild()`. This has proven too limited for a couple reasons:

* It can't handle modules such as eventing that include their own
  Go-based code-generation tools that must be built before `go mod tidy`
  can be run.
* It doesn't do anything for library modules that have no `GoModBuild()`
  targets.

Instead, now we introduce a `GoModTidySetup()` macro that explicitly
declares each Go module directory, as well as a list of library modules
that should be tidied.

Also, remove a couple old workarounds that we needed with earlier Go
versions.

Change-Id: I2b329796dbb6c5a61f8044bb4fc798e9286e23c5
Reviewed-on: https://review.couchbase.org/c/tlm/+/224887
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>