MB-64581: cb_install_code() needs to install to default component
The new `cb_install_code()` function mistakenly uses `ALL_COMPONENTS`
when a specific component was not specified. This results in some normal
build artifacts being re-installed during standalone-package
installation, which is wasteful and occasionally led to race condition
errors. Drop the `ALL_COMPONENTS` parameter so that the install-time
code is only added to the default component.

Change-Id: Ifa9b4fe324deaee4705fb5bacac6f935844c9cfc
Reviewed-on: https://review.couchbase.org/c/tlm/+/230614
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>