MB-66289: Set -march=x86-64-v3 for g++ and ...
... -mavx2 -msse4.2 compile options for clang. The options are
supported even on architectures which don't support them (using
check_c_compiler_flag does not fail)

Folly is rebuilt with AVX2 support due to link problems trying
to build just the server with these options.

Due to the fact that kv_engine CV runs a mix of gcc and clang
(in the asan/ubsan build) both compiler options must be specified
in the same patch

This commit was originally submitted as 7b4ff098, and then reverted
while upgrading the Perf test clusters to newer hardware.

Change-Id: I23c41a49e930cd205ce37a3f6af5cb1748979f56
Reviewed-on: https://review.couchbase.org/c/tlm/+/230939
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>