MB-68561: Add suppression for deadlock...
... in ExpiredItemsCallback::callback

The deadlock is caused by lock order inversion with the VBucket
state lock and magma's serialization for rollback and compaction.

Compaction would acquire Magma's serialization lock first and
then the VBucket state lock when calling processExpiredItem from
ExpiredItemsCallback::callback

Rollback on the other hand acquire the VBucket state lock
first before calling magma's rollback which acquire the
serialization lock.

This is not a problem in the current code as we (kv) refuse
to start one of the tasks if the other one is running, and
internal compaction in magma is stopped before we try to
execute rollback.

Change-Id: I900091280cc70d8619066242d61b256e8519ce6e
Reviewed-on: https://review.couchbase.org/c/tlm/+/234281
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>