MB-68973: Force rebar3 to use cache in binary directory
`rebar3` normally caches downloaded packages in `~/.cache/rebar3` which
persists between builds. However it has a bug that can cause it to use
an older cached version of an Erlang plugin rather than downloading the
correct newer version. Most straightforward way to avoid this is to tell
it to use a cache in the CMake binary directory, which will always be
created fresh on production builds and will be relatively short-lived
for developer builds. Fortunately the amount of data downloaded to this
cache is quite small.

Change-Id: I794dc2d5f9a2b461e51776d6daa96eb3f73a7f77
Reviewed-on: https://review.couchbase.org/c/tlm/+/235269
Tested-by: Chris Hillery <ceej@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Blair Watt <blair.watt@couchbase.com>