CBD-6507: Fix failing black duck scans
Since rebuilding some dependencies on the centos based single-linux,
the libraries are in a directory called `lib64` rather than `lib`.
This change ensures we detect libs in both dir names.

Although we have a similar mix of lib and lib64 dirs on other
releases, we haven't come across this issue on those because their
versions of the FindCouchbaseGlog and FindCouchbasePrometheus modules
set this same global property prior to dependencies with lib64 dirs
being reached. On Neo we do not set this property in any of the Find*
modules, so started encountering this error as soon as we stepped
away from distro-specific dependency builds.

Change-Id: Ic56c63ad09ba4c460ab2ac47a245f8d4684411ee
Reviewed-on: https://review.couchbase.org/c/tlm/+/236463
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker